Avoid dynamic ArrayList growth in HitGroup when number of hits is known
